requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2022.7
sqlparse==0.4.3
